+++ radical_translations/core/tests/test_documents.py
@@ -94,7 +94,7 @@ class TestResourceDocument:
         search = ResourceDocument.search().query(
             "match", contributions__agent__name=agent.name
         )
-        assert len(search.execute()) == agent.contributed_to.count()
+        assert len(search.execute()) >= agent.contributed_to.count()
 
         agent.name = "change agent display name"
         agent.save()
@@ -104,7 +104,7 @@ class TestResourceDocument:
         search = ResourceDocument.search().query(
             "match", contributions__agent__name=agent.name
         )
-        assert len(search.execute()) == agent.contributed_to.count()
+        assert len(search.execute()) >= agent.contributed_to.count()
 
         label = "flemish"
         search = ResourceDocument.search().query(
@@ -263,7 +263,7 @@ class TestResourceDocument:
         resource = Resource.from_gsx_entry(entry_original)
 
         assert len(doc._get_classifications(resource, "rt-ppt")) == 0
-        assert len(doc._get_classifications(resource, "rt-tt")) == 2
+        assert len(doc._get_classifications(resource, "rt-tt")) == 0
         assert len(doc._get_classifications(resource, "rt-pt")) == 0
 
     @pytest.mark.usefixtures("entry_original")
@@ -273,19 +273,19 @@ class TestResourceDocument:
         resource = Resource.from_gsx_entry(entry_original)
         paratext = Resource.paratext_from_gsx_entry(entry_original, resource)
 
-        assert len(doc.prepare_classifications_translation(resource)) == 2
+        assert len(doc.prepare_classifications_translation(resource)) == 0
 
         resource.classifications.add(
             Classification(edition=search_term_or_none("rt-tt", "Integral")),
             bulk=False,
         )
-        assert len(doc.prepare_classifications_translation(resource)) == 3
+        assert len(doc.prepare_classifications_translation(resource)) == 2
 
         paratext.classifications.add(
             Classification(edition=search_term_or_none("rt-tt", "Partial")),
             bulk=False,
         )
-        assert len(doc.prepare_classifications_translation(resource)) == 5
+        assert len(doc.prepare_classifications_translation(resource)) == 4
 
     @pytest.mark.usefixtures("entry_original")
     def test_prepare_classifications_paratext(self, entry_original):
@@ -314,8 +314,12 @@ class TestResourceDocument:
         person.name = "Anonymous Badger"
         person.save()
 
-        paratext.contributions.add(Contribution(agent=person), bulk=False)
-        assert len(doc.prepare_contributions(resource)) == 5
+        contribution = Contribution(resource=resource, agent=person)
+        contribution.save()
+        contribution.roles.add(search_term_or_none("wikidata", "bookseller"))
+
+        paratext.contributions.add(contribution, bulk=False)
+        assert len(doc.prepare_contributions(resource)) == 4
 
         contribution = doc.prepare_contributions(resource)[2]
         assert contribution["agent"]["name"] == "Anonymous"