+++ radical_translations/events/models.py
@@ -1,14 +1,18 @@
 from typing import Dict, Optional
 
+from django.conf import settings
 from django.db import models
-from geonames_place.models import Place
 from model_utils.models import TimeStampedModel
+
+from geonames_place.models import Place
 from radical_translations.core.models import Resource
 from radical_translations.utils.models import (
     Date,
     EditorialClassificationModel,
+    date_to_dict,
     get_geonames_place_from_gsx_place,
     get_gsx_entry_value,
+    place_to_dict_value,
 )
 
 # These models are based on the BIBFRAME 2.0 Event model
@@ -48,6 +52,21 @@ class Event(TimeStampedModel, EditorialClassificationModel):
     def __str__(self) -> str:
         return f"{self.date}: {self.title}"
 
+    def get_classification(self) -> str:
+        return "; ".join([c.label for c in self.classification.all()])
+
+    get_classification.short_description = "Classification"  # type: ignore
+
+    def to_dict(self) -> Dict:
+        return {
+            "id": self.id,
+            **date_to_dict(self.date),
+            "place": place_to_dict_value(self.place),
+            "related_to": f"{settings.EXPORT_MULTIVALUE_SEPARATOR} ".join(
+                [r.to_dict_value() for r in self.related_to.all()]
+            ),
+        }
+
     @staticmethod
     def from_gsx_entry(entry: Dict[str, Dict[str, str]]) -> Optional["Event"]:
         """Gets or creates a new `Event` from a Google Spreadsheet dictionary
@@ -66,5 +85,4 @@ class Event(TimeStampedModel, EditorialClassificationModel):
         )
 
         event, _ = Event.objects.get_or_create(title=title, date=date, place=place)
-
         return event