+++ radical_translations/events/serializers.py
@@ -1,0 +1,8 @@
+from django_elasticsearch_dsl_drf.serializers import DocumentSerializer
+
+from radical_translations.events.documents import EventDocument
+
+
+class EventDocumentSerializer(DocumentSerializer):
+    class Meta:
+        document = EventDocument